fix(gpu): signed EIP-2200 refund counter, EIP-3529 cap at dispatcher
The EIP-2200 SSTORE accounting subtracts SSTORE_REFUND when a previously
credited refund must be reversed (e.g. orig != 0, cur == 0). Both kernels
and the kernel CPU interpreter stored the refund counter as unsigned
(ulong/uint64), so that subtraction underflowed to ~2^64 and corrupted
gas_used after the EIP-3529 cap was applied.

This change:

* Promotes the refund counter to signed int64 in all three paths:
  Metal (long), CUDA (long long), kernel CPU (int64_t).
* Updates the wire-layout TxOutput.gas_refund to signed in
  evm_kernel.metal, evm_kernel.cu, kernel/evm_kernel_host.hpp, and
  cuda/evm_kernel_host.hpp. Same width (8 bytes) — static_assert(
  sizeof(TxOutput) == 32) still holds.
* Adds gas_refund to kernel::TxResult and propagates from Metal
  evm_kernel_host.mm (CUDA host already did so via go.gas_refund).
* Adds InterpreterResult::gas_refund (default 0) for the CPU
  interpreter path.
* Applies the EIP-3529 cap (refund <= gas_used / 5; final gas =
  gas_used - capped_refund; never increases gas_used) at the
  dispatcher in gpu_dispatch.cpp when assembling BlockResult — for
  CPU sequential, CPU parallel kernel path, GPU Metal, and GPU CUDA.
  Negative refund counters floor at 0 (refund never increases gas).

ABI: gas_refund is internal to the kernel<->host wire format. The
external Go ABI (go_bridge.h CGpuBlockResultV2) exposes only the
post-cap gas_used, so EVM_GPU_ABI_VERSION stays at 3.

Test: test/unittests/refund_test.cpp drives four scenarios through
Backend::CPU_Sequential (which routes bytecode through the kernel CPU
interpreter, the same one Metal and CUDA emulate):
  1. CreateSlotNoRefund — orig=0, cur=0, nv!=0 → SET, no refund
  2. ResetSlotCreditsRefund — A->0 credits SSTORE_SET-NOOP, cap moot
  3. SignedCounterNoUnderflow — A->0->B->0 cycle: signed counter
     reaches the right final value; pre-fix this exhibited underflow
  4. RefundCapAppliedAtDispatcher — refund > gas_used/5 → clamped

All four pass. Existing 4-way parity test (133 vectors across
CPU_Sequential, CPU_Parallel, GPU_Metal) still passes.

+++ lib/evm/gpu/kernel/evm_kernel.metal
@@ -351,7 +351,7 @@ struct TxInput {
     ulong gas_limit; uint256 caller; uint256 address; uint256 value;
 };
 struct TxOutput {
-    uint status; ulong gas_used; ulong gas_refund; uint output_size;
+    uint status; ulong gas_used; long gas_refund; uint output_size;
 };
 struct StorageEntry { uint256 key; uint256 value; };
 
@@ -415,11 +415,14 @@ static inline bool original_value_lookup(thread OriginalEntry* o,uint n,uint256 
 static inline void original_value_record(thread OriginalEntry* o,thread uint& n,uint256 s,uint256 v){
     for(uint i=0;i<n;++i)if(o[i].valid&&u256_eq(o[i].key,s))return;
     if(n<MAX_STORAGE_PER_TX){o[n].key=s;o[n].value=v;o[n].valid=true;n++;}}
-static inline ulong sstore_gas_eip2200(uint256 orig,uint256 cur,uint256 nv,thread ulong& rc){
+// Signed `rc` so refund subtraction can dip below zero per EIP-2200 ordering
+// (e.g. orig!=0, cur==0 then write nv==0 cancels a prior credit). Cap to
+// gas_used/5 happens at the dispatcher (EIP-3529) once execution finishes.
+static inline ulong sstore_gas_eip2200(uint256 orig,uint256 cur,uint256 nv,thread long& rc){
     if(u256_eq(nv,cur))return GAS_SSTORE_NOOP;
-    if(u256_eq(orig,cur)){if(u256_iszero(orig))return GAS_SSTORE_SET;if(u256_iszero(nv))rc+=GAS_SSTORE_REFUND;return GAS_SSTORE_RESET;}
-    if(!u256_iszero(orig)){if(u256_iszero(cur))rc-=GAS_SSTORE_REFUND;else if(u256_iszero(nv))rc+=GAS_SSTORE_REFUND;}
-    if(u256_eq(nv,orig)){if(u256_iszero(orig))rc+=GAS_SSTORE_SET-GAS_SSTORE_NOOP;else rc+=GAS_SSTORE_RESET-GAS_SSTORE_NOOP;}
+    if(u256_eq(orig,cur)){if(u256_iszero(orig))return GAS_SSTORE_SET;if(u256_iszero(nv))rc+=(long)GAS_SSTORE_REFUND;return GAS_SSTORE_RESET;}
+    if(!u256_iszero(orig)){if(u256_iszero(cur))rc-=(long)GAS_SSTORE_REFUND;else if(u256_iszero(nv))rc+=(long)GAS_SSTORE_REFUND;}
+    if(u256_eq(nv,orig)){if(u256_iszero(orig))rc+=(long)(GAS_SSTORE_SET-GAS_SSTORE_NOOP);else rc+=(long)(GAS_SSTORE_RESET-GAS_SSTORE_NOOP);}
     return GAS_SSTORE_NOOP;
 }
 
@@ -499,7 +502,7 @@ kernel void evm_execute(
     uint256 stack[32];
     uint sp = 0;
     ulong gas = inp.gas_limit;
-    ulong refund_counter = 0;
+    long  refund_counter = 0;
     uint pc = 0;
     uint mem_size = 0;
     ulong gas_start = gas;
